Split POTRF into multiple files; rename main test program
