+++ examples/potrf/core_plgsy.h
@@ -1,3 +1,5 @@
+#pragma once
+
 /**
  *
  * @file core_dplgsy.c
